Refactor custom middleware into its own directory
The validation and mapping middleware should be separate from the route
code to not mix up concerns.

OHM-973
